requirements.txt
certifi==2018.11.29
chardet==3.0.4
get==2018.11.19
idna==2.7
numpy==1.15.4
Pillow==5.3.0
post==2018.11.20
public==2018.11.20
PyQt5==5.11.3
PyQt5-sip==4.19.13
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
urllib3==1.24.1
